+++ test/result_test.dart
@@ -25,17 +25,6 @@ void main() {
       expect(result, value);
     });
 
-    test("executes the success branch for null success value", () {
-      const int? value = null;
-      const sut = Result<int, int?>.success(value);
-      final handler = _ResultHandler<int?, int?, int?>();
-
-      final result =
-          sut.iif(success: handler.success, failure: handler.failure);
-
-      expect(result, null);
-    });
-
     test("throws failure as is", () {
       final failureObject = Exception();
       final sut = Result<int, Exception>.failure(failureObject);